requirements.txt
Pillow==9.2.0
pyperclip==1.8.2
PySimpleGUI==4.60.3
requests~=2.28.1